fix api where magic number was expected to be long elsewhere but being passed as an int
git-svn-id: http://svn.eol.ucar.edu/svn/rdss/trunk/zebra/source@8275 3f9b6c58-9310-455a-aec6-8b30269ab333

+++ src/lib/BlockFile.hh
@@ -93,7 +93,8 @@ public:
 	// file is opened or a new file is created.  Pass BF_CREATE to
 	// explicitly re-create an existing file.
 	//
-	BlockFile (const char *path, int app_magic = 0, int flags = 0);
+	BlockFile (const char *path, unsigned long app_magic = 0, 
+		   int flags = 0);
 
 	// Destructor
